feat: add CategoryRepository and refactor LabRepository to use it for category management
