+++ app/domain/university/repository.py
@@ -1,6 +1,5 @@
 from app.common.base_repository import BaseRepository
 from app.domain.university.model import University
-from app.domain.university.schema import UniversityCreateSchema, UniversityOutSchema
 
 
 class UniversityRepository(BaseRepository[University]):
